ai_service/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
cryptography==42.0.8
pydantic-settings==2.3.0
